Fixed issues related to comment -- see below
can you move the buttons so they are more centered and larger, and have the "Close" button next to the Submit. I couldn't find the close button until I searched for it. Submit button should be green. Close button could be red and labelled "Cancel" until the testimonial is submitted.